LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/embedded-content/bfcache/embedded-type-only-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

Harness Error (TIMEOUT), message = null

TIMEOUT Page with <embed type=text/html src=undefined> Test timed out
NOTRUN Page with <object type=text/html data=undefined>
